Task #12 - Refactor TokenHandler Fixes #12
- moved the RSA key generation to AuthenticationMechanism
- moved the token generation and parsing into the newly renamed TokenIdentityStore
- removed the unneeded TokenHandler
- reworked the TokenCredential to just be a holder for the auth header
