+++ security/src/main/java/com/github/nderwin/micro/playground/security/control/TokenCredential.java
@@ -1,52 +1,21 @@
 package com.github.nderwin.micro.playground.security.control;
 
-import io.jsonwebtoken.Claims;
-import java.util.Date;
-import java.util.List;
-import javax.json.Json;
-import javax.json.JsonArrayBuilder;
-import javax.json.JsonObject;
-import javax.json.JsonObjectBuilder;
 import javax.security.enterprise.credential.Credential;
 
 public class TokenCredential implements Credential {
 
-    private final Claims claims;
+    private final String token;
 
-    public TokenCredential(final Claims claims) {
-        if (null == claims) {
-            throw new IllegalArgumentException("Claims must not be null");
+    public TokenCredential(final String token) {
+        if (null == token) {
+            throw new IllegalArgumentException("Token must not be null");
         }
         
-        this.claims = claims;
+        this.token = token;
     }
 
-    public String getSubject() {
-        return this.claims.getSubject();
+    public String getToken() {
+        return token;
     }
 
-    public List getScope() {
-        return this.claims.get("scope", List.class);
-    }
-
-    public Date getExpirationDate() {
-        if (null != this.claims.getExpiration()) {
-            return new Date(this.claims.getExpiration().getTime());
-        }
-        
-        return null;
-    }
-
-    public JsonObject toJson() {
-        JsonObjectBuilder job = Json.createObjectBuilder();
-        job.add("subject", getSubject());
-
-        JsonArrayBuilder jab = Json.createArrayBuilder();
-        getScope().forEach((o) -> {
-            jab.add(o.toString());
-        });
-        job.add("scope", jab.build());
-        
-        return job.build();
-    }
 }
